backend/requirements.txt
Django==6.0.1
djangorestframework
djangorestframework-simplejwt
pdfplumber
python-docx
django-cors-headers
